oz-invoice-hybrid/backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.5
requests==2.31.0
marshmallow==3.21.1
PyJWT==2.8.0
gunicorn==22.0.0
pytest==8.0.1
pytest-flask==1.3.0
